1、算牛相关。缺牛型提示。

+++ Classes/GamePlayScene.h
@@ -157,7 +157,7 @@ private:
     CDMRecordObject *m_recordObject;
 #endif
 	bool m_bGameStart;
-	vector<unsigned long long > m_testID;
+	vector<int> m_testID;
 };
